model tests for users db
